+++ crd-from-oas/pkg/generator/templates.go
@@ -256,6 +256,15 @@ func (obj *{{.EntityName}}) Get{{.RootRefAccessorEntityName}}Ref() {{.RootRefTyp
 	return obj.Spec.{{.RootRefDependency.FieldName}}
 }
 {{- end}}
+
+// GetParentRef returns the reference to the parent entity.
+func (obj *{{.EntityName}}) GetParentRef() {{.RootRefTypeName}} {
+{{- if gt (len .RootRefAccessorEntityName) (len .RootRefDependency.EntityName)}}
+	return obj.Get{{.RootRefAccessorEntityName}}Ref()
+{{- else}}
+	return obj.Get{{.RootRefDependency.EntityName}}Ref()
+{{- end}}
+}
 {{- end}}
 {{- if .IsReconcilerRoot}}
 
@@ -972,7 +981,9 @@ func update{{.Entity}}(
 		return CantPerformOperationWithoutParentIDError{Entity: obj, Parent: "{{.EntityName}}", Op: UpdateOp}
 	}
 {{- end}}
+{{- if not .UpdateOmitsEntityID}}
 	id := obj.GetKonnectStatus().GetKonnectID()
+{{- end}}
 	{{- if .NeedsClient}}
 	req, err := obj.To{{.UpdateReqType}}(ctx, cl)
 	{{- else}}
@@ -1004,6 +1015,9 @@ func update{{.Entity}}(
 		{{.EntityIDField}}: id,
 		{{.UpdateBodyField}}: {{if .UpdateReqBodyPointer}}req{{else}}*req{{end}},
 	})
+{{- else if .UpdateOmitsEntityID}}
+
+	_, err = sdk.{{.UpdateSDKMethod}}(ctx, {{(index .Parents 0).VarName}}, {{if .UpdateReqBodyPointer}}req{{else}}*req{{end}})
 {{- else}}
 
 	_, err = sdk.{{.UpdateSDKMethod}}(ctx, id, {{if .UpdateReqBodyPointer}}req{{else}}*req{{end}})
@@ -1019,8 +1033,9 @@ func update{{.Entity}}(
 
 // opsDeleteFuncTemplate renders a single delete<Entity> function body.
 // It is concatenated after the file header and any create/update functions.
-// Single-parent SDK delete methods use positional arguments.
 // Multi-parent SDK delete methods use a fully-wrapped request struct.
+// Parent-scoped singletons omit the entity ID (not in path).
+// Single-parent SDK delete methods use positional arguments.
 // Optional query parameters (e.g. force) are passed as nil.
 const opsDeleteFuncTemplate = `
 func delete{{.Entity}}(
@@ -1034,7 +1049,9 @@ func delete{{.Entity}}(
 		return CantPerformOperationWithoutParentIDError{Entity: obj, Parent: "{{.EntityName}}", Op: DeleteOp}
 	}
 {{- end}}
+{{- if not .DeleteOmitsEntityID}}
 	id := obj.GetKonnectStatus().GetKonnectID()
+{{- end}}
 {{- if .DeleteFullyWrapped}}
 
 	_, err := sdk.{{.DeleteSDKMethod}}(ctx, sdkkonnectops.{{.DeleteWrappedType}}{
@@ -1043,6 +1060,9 @@ func delete{{.Entity}}(
 		{{- end}}
 		{{.DeleteEntityIDField}}: id,
 	})
+{{- else if .DeleteOmitsEntityID}}
+
+	_, err := sdk.{{.DeleteSDKMethod}}(ctx, {{(index .Parents 0).VarName}}{{range .DeleteNilArgs}}, nil{{end}})
 {{- else if .Parents}}
 
 	_, err := sdk.{{.DeleteSDKMethod}}(ctx, {{(index .Parents 0).VarName}}, id{{range .DeleteNilArgs}}, nil{{end}})
@@ -1496,7 +1516,6 @@ const groupVersionInfoTemplate = sharedGeneratedFilePreamble + `
 package {{.APIVersion}}
 
 import (
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 )
@@ -1526,6 +1545,20 @@ func Resource(resource string) schema.GroupResource {
 }
 
 func addKnownTypes(scheme *runtime.Scheme) error {
+	return addKnownTypesGenerated(scheme)
+}
+`
+
+const groupVersionInfoGeneratedTemplate = sharedGeneratedFilePreamble + `
+
+package {{.APIVersion}}
+
+import (
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+)
+
+func addKnownTypesGenerated(scheme *runtime.Scheme) error {
 	scheme.AddKnownTypes(GroupVersion,
 {{- range .EntityNames}}
 		&{{.}}{},